[KAIZEN-0] flytte SafService ut av legacy package
