+++ web/src/main/java/no/nav/modiapersonoversikt/service/saker/SakerServiceImpl.kt
@@ -3,7 +3,7 @@
 package no.nav.modiapersonoversikt.service.saker
 
 import no.nav.arena.services.sakvedtakservice.SakVedtakPortType
-import no.nav.modiapersonoversikt.legacy.sak.service.saf.SafService
+import no.nav.modiapersonoversikt.service.saf.SafService
 import no.nav.modiapersonoversikt.service.enhetligkodeverk.EnhetligKodeverk
 import no.nav.modiapersonoversikt.service.enhetligkodeverk.KodeverkConfig
 import no.nav.modiapersonoversikt.service.saker.kilder.*
